pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qclab"
version = "1.0.0"
description = "Reduced, benchmarked, and interpretable Qiskit workflows for early-universe research models."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [{ name = "John Blattner" }]
keywords = [
  "quantum cosmology",
  "qiskit",
  "early universe",
  "scientific computing",
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
  "matplotlib>=3.8",
  "numpy>=1.26",
  "pandas>=2.2",
  "PyYAML>=6.0",
  "qiskit>=2.2,<3",
  "qiskit-aer>=0.17,<1",
  "qiskit-ibm-runtime>=0.46,<1",
  "scipy>=1.13",
  "sympy>=1.13",
]

[project.optional-dependencies]
dev = ["pytest>=8"]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
